==> Sonoran-scripts/04.1_data-wrangling_subplot.R <==
# Created: 2025-01-31
# Last updated: 2025-10-10

# Purpose: Create clean data table for subplot data (Count), with corrected and standardized species information,
#   and monitoring and plot information, and correct SpeciesSeeded column based on each site-specific
#     seed mix and plot.
#   Essentially, add corrected metadata from 01.R and 02.R to subplot data,
#     correct SpeciesSeeded column, and add other grouping columns (PlantSource, PlantSource2, Weedy, PlotMix_Climate).


library(readxl)
library(tidyverse)

# Load data ---------------------------------------------------------------

subplot.se.raw <- read_xlsx("Sonoran-data/raw/Sonoran_2023-09-15_Master 1.0 Germination Data_LO.xlsx", 
                            sheet = "SonoranSE_Subplots_LO")
subplot.cen.raw <- read_xlsx("Sonoran-data/raw/Sonoran_2023-09-15_Master 1.0 Germination Data_LO.xlsx", 
                             sheet = "SonoranCentral_Subplots_LO")
species.in <- read_csv("Sonoran-data/cleaned/01_subplot_species-list_location-independent_clean.csv")
species.de <- read_csv("Sonoran-data/cleaned/01_subplot_species-list_location-dependent_clean.csv")
mix <- read_xlsx("Sonoran-data/raw/from-Master_seed-mix_LO_Sonoran.xlsx", sheet = "with-site_R") %>% 
  filter(Site %in% c("SRER", "Patagonia", "Preserve", "Pleasant", "SCC", "Roosevelt"))
monitor.info <- read_csv("Sonoran-data/cleaned/02_corrected-monitoring-info_clean.csv")
monitor.wrong <- read_csv("Sonoran-data/data-wrangling-intermediate/02_subplot-wrong-monitor-events.csv")
monitor.fixed <- read_csv("Sonoran-data/data-wrangling-intermediate/02_subplot-wrong-monitor-events-corrected.csv")
monitor.site <- read_csv("Sonoran-data/cleaned/02_SiteDateID_clean.csv")
monitor.siteplot <- read_csv("Sonoran-data/cleaned/02_SitePlotID_clean.csv")


# Organize columns --------------------------------------------------------

# Convert cols to character and combine all Sonoran Desert sites 
subplot.se <- subplot.se.raw %>% 
  mutate(across(everything(), as.character))
subplot.cen <- subplot.cen.raw %>% 
  mutate(across(everything(), as.character))
subplot <- bind_rows(subplot.se, subplot.cen)

# Narrow down subplot.raw columns
subplot <- subplot %>%
  select(-Recorder_Initials, -Functional_Group, -`Certainty_of_ID(1-3)`, -Notes) %>%
  mutate(raw.row = 1:nrow(subplot)) %>% # row number is to be able to easily refer back to the raw data and excluded columns if needed
  rename(CodeOriginal = Species_Code,
         Count = Seedling_Count,
         Height = Average_Height_mm,
         SpeciesSeeded = `Seeded(Yes/No)`,
         PlotMix = Seed_Mix)

# Add Region col
subplot <- subplot %>%
  mutate(Region = case_when(
    str_detect(subplot$Site, c("SRER|Patagonia")) ~ "Sonoran SE",
    str_detect(subplot$Site, c("Preserve|SCC|Roosevelt|Pleasant")) ~ "Sonoran Central"))

# Convert character "NA" to logical NA
subplot[subplot == "NA"] <- NA

# Check for NAs
apply(subplot, 2, anyNA) # check all columns for NAs; NAs in Count, Height, SpeciesSeeded are expected

# Remove incomplete monitoring events 
#   2022-10-06 at Pleasant and 2022-10-05 at Roosevelt (not all subplots were sampled)
subplot <- subplot %>% 
  filter(!Date_Monitored %in% c("2022-10-06", "2022-10-05"))

# Convert Date_Seeded and Date_Monitored to date cols, and Plot to numeric
subplot <- subplot %>%
  mutate(Date_Seeded = as.Date(Date_Seeded),
         Date_Monitored = as.Date(Date_Monitored),
         Plot = as.numeric(Plot)) 



# Add species info for location-dependent ---------------------------------

# Separate out location-dependent observations
subplot.de <- subplot %>%
  filter(CodeOriginal %in% species.de$CodeOriginal)

# Add species info
subplot.de <- left_join(subplot.de, species.de)

# Check for NA codes
filter(subplot.de, is.na(Code)) # no NAs
apply(subplot.de, 2, anyNA) # NAs for Count, Height, SpeciesSeeded are expected



# Add species info for location-independent codes -------------------------

subplot.in <- subplot %>%
  filter(CodeOriginal %in% species.in$CodeOriginal)

subplot.in <- left_join(subplot.in, species.in)

# Check for NA codes
filter(subplot.in, is.na(Code)) # no NAs
apply(subplot.in, 2, anyNA) # NAs for Count, Height, SpeciesSeeded inherent to data



# Combine location-dependent and independent ------------------------------

subplot <- bind_rows(subplot.in, subplot.de) %>%
  arrange(raw.row)

# Check that there the same number of observations as the original subplot data 
#   (Sonoran sites only, without incomplete 10/5 and 10/6 sampling)
nrow(subplot) == (nrow(subplot.cen.raw) + nrow(subplot.se.raw) - 20)



# Check if Introduced plants were marked as Seeded ------------------------

subplot.inva <- subplot %>%
  filter(Native == "Introduced") %>%
  select(Code, SpeciesSeeded, Name, Native) %>%
  distinct(.keep_all = TRUE)
unique(subplot.inva$SpeciesSeeded) # something is mislabeled; no introduced species were seeded
subplot.inva %>%
  filter(SpeciesSeeded == "Yes")

# Fix ERCU2 and ERCI6 - mark all observations as "No" for Seeded col
subplot$SpeciesSeeded[subplot$Name == "Eragrostis curvula"] <- "No"
subplot$SpeciesSeeded[subplot$Name == "Erodium cicutarium"] <- "No"



# Correct monitoring info -------------------------------------------------

# Separate out monitoring info from rest of subplot data
subplot.monitor <- subplot %>%
  select(Region, Site, Date_Seeded, Date_Monitored, Plot, Treatment, PlotMix, raw.row)

# Find raw.row number of events that need to be fixed
#   some events that need to be fixed have multiple rows in subplot data
wrong.raw.row <- left_join(monitor.wrong, subplot.monitor)

# Attach raw.row to corrected monitoring info
monitor.assign <- data.frame(SiteDatePlotID = wrong.raw.row$SiteDatePlotID)
monitor.assign <- left_join(monitor.assign, monitor.fixed)
#   wrong.raw.row and monitor.assign have the same row order (rows already correspond)
monitor.assign$raw.row <- wrong.raw.row$raw.row

# Separate monitor info that doesn't need to be fixed and add SiteDatePlotID
subplot.monitor <- subplot.monitor %>%
  filter(!raw.row %in% monitor.assign$raw.row) %>%
  left_join(monitor.info)
subplot.monitor %>%
  filter(is.na(SiteDatePlotID)) # all assigned SiteDatePlotID

# Add corrected monitor info for complete list
subplot.monitor <- bind_rows(subplot.monitor, monitor.assign) %>%
  arrange(SiteDatePlotID)

# Check for matching lengths (Sonoran sites only)
nrow(subplot) == (nrow(subplot.cen.raw) + nrow(subplot.se.raw) - 20)


# Attach correct monitoring info to subplot data
subplot <- subplot[, -c(1:6)] # remove incorrect info
subplot <- left_join(subplot, subplot.monitor) %>% # add correct info
  arrange(raw.row)

# Reorder columns
subplot <- subplot %>%
  select(Region, Site, Date_Seeded, Date_Monitored, Plot, Treatment, PlotMix, SiteDatePlotID,
         CodeOriginal, Code, Name, Native, Duration, Lifeform, SpeciesSeeded,
         Count, Height, raw.row)


# Check all cols for NAs
apply(subplot, 2, anyNA) # NA in Count, Height, SpeciesSeeded is okay



# Check for all SiteDatePlotIDs
nrow(monitor.info) # 1152 IDs
length(unique(subplot$SiteDatePlotID)) == nrow(monitor.info)


# Add SiteDateID and SitePlotID now that subplot data is complete
subplot <- left_join(subplot, monitor.site) %>%
  left_join(monitor.siteplot) %>% 
  arrange(raw.row)

# Reorder columns again
subplot <- subplot %>%
  select(Region, Site, Date_Seeded, Date_Monitored, SiteDateID, Plot, Treatment, PlotMix, SitePlotID,
         SiteDatePlotID, CodeOriginal, Code, Name, Native, Duration, Lifeform, SpeciesSeeded,
         Count, Height, raw.row)

# Save intermediate subplot df, which has:
#   Correct monitoring info
#   Correct Native, Duration, Lifeform
#   Incorrect SeededSpecies (not yet addressed)
subplot1 <- subplot



# Correct SpeciesSeeded column --------------------------------------------

# Address seeded species codes not in mix

# Examine possible SpeciesSeeded values
#   then deal with each separately in own section
unique(subplot$SpeciesSeeded)


# Species seeded but not listed in mix by code
subplot.seeded <- subplot %>%
  filter(SpeciesSeeded %in% c("Yes", "Y", "y", "Yes?"))

setdiff(unique(subplot.seeded$Code), unique(mix$CodeOriginal)) # BOUSPP.SRER, BOER4

species.seeded.not.in.mix <- subplot.seeded %>% 
  filter(Code %in% c("BOUSPP.SRER", "BOER4")) %>% 
  select(Site, Region, PlotMix, CodeOriginal, Code, Name, Native, Duration, Lifeform,
         SpeciesSeeded) 
species.seeded.not.in.mix

# Change "Yes?" to "No", as it is an unknown
species.seeded.not.in.mix$SpeciesSeeded[species.seeded.not.in.mix$Code == "BOUSPP.SRER"] <- "No"
subplot$SpeciesSeeded[subplot$Code == "BOUSPP.SRER"] <- "No"

# Change BOER4 to "No", as BOER4 was not seeded at SRER
species.seeded.not.in.mix$SpeciesSeeded[species.seeded.not.in.mix$Code == "BOER4"] <- "No"
subplot$SpeciesSeeded[subplot$Code == "BOER4"] <- "No"


# Species seeded and in (a) mix
#   Need to look at manually because mixes are site-specific
species.seeded.in.mix <- subplot.seeded %>%
  filter(SpeciesSeeded %in% c("Yes", "Y", "y")) %>%
  select(Site, Region, PlotMix, CodeOriginal, Code, Name, Native, Duration, Lifeform, 
         SpeciesSeeded) %>%
  distinct(.keep_all = TRUE) %>%
  arrange(Name) %>%
  arrange(Site) %>%
  arrange(Region)

# OUTPUT: create list of species marked seeded and in mix
write_csv(species.seeded.in.mix,
          file = "Sonoran-data/data-wrangling-intermediate/04.1a_output1_seeded-in-mix_subplot.csv")

# EDITED: manually review and fix SpeciesSeeded status
#   Standardize responses so all are in format "Yes"
#   Incorrect ones corrected
species.seeded.in.mix <- read_xlsx("Sonoran-data/data-wrangling-intermediate/04.1b_edited1_corrected-seeded-in-mix_subplot.xlsx")

# Remove duplicates that resulted from standardizing "Yes"
species.seeded.in.mix <- species.seeded.in.mix %>%
  distinct(.keep_all = TRUE)


# Species marked Unknown ("?")
species.unk.seeded <- subplot %>%
  filter(SpeciesSeeded == "?") %>%
  select(Site, Code, Name, SpeciesSeeded) 
species.unk.seeded

# Change "?" to "No", as this is an unknown grass
species.unk.seeded$SpeciesSeeded[species.unk.seeded$Code == "UNGR_S_2.SRER"] <- "No"
subplot$SpeciesSeeded[subplot$Code == "UNGR_S_2.SRER"] <- "No"



# Species marked NA
species.na.seeded <- subplot %>%
  filter(is.na(SpeciesSeeded)) %>%
  select(Site, Region, PlotMix, CodeOriginal, Code, Name, Native, Duration, Lifeform,
         SpeciesSeeded) %>%
  distinct(.keep_all = TRUE) %>%
  arrange(Name) %>%
  arrange(Site) %>%
  arrange(Region)

# OUTPUT: create list of species of NA seeding status
write_csv(species.na.seeded,
          file = "Sonoran-data/data-wrangling-intermediate/04.1a_output2_NA-seeded_subplot.csv")

# EDITED: manually review and fix SpeciesSeeded status
#   Unknowns marked as not seeded.
#   0 code marked as "0" (no plants were present)
#  Seeded status changed if identified to species level and it was seeded.
species.na.seeded <- read_xlsx("Sonoran-data/data-wrangling-intermediate/04.1b_edited2_corrected-NA-seeded_subplot.xlsx")



# Species marked not seeded
subplot.no <- subplot %>%
  filter(SpeciesSeeded %in% c("No", "N"))

species.no.seeded <- subplot.no %>%
  select(Site, Region, PlotMix, CodeOriginal, Code, Name, Native, Duration, Lifeform,
         SpeciesSeeded) %>%
  distinct(.keep_all = TRUE) %>%
  arrange(Name) %>%
  arrange(Site) %>%
  arrange(Region)

# OUTPUT: create list of species marked not seeded
write_csv(species.no.seeded,
          file = "Sonoran-data/data-wrangling-intermediate/04.1a_output3_not-seeded_subplot.csv")

# EDITED: manually review and fix SpeciesSeeded status
#   Standardize responses so all are in format "Yes"
#   Incorrect ones corrected
species.no.seeded <- read_xlsx("Sonoran-data/data-wrangling-intermediate/04.1b_edited3_corrected-not-seeded_subplot.xlsx")

# Remove duplicates that resulted from standardizing "No"
species.no.seeded <- species.no.seeded %>%
  distinct(.keep_all = TRUE)


# Compile list of correct SpeciesSeeded
#   Correct SpeciesSeeded status based on PlotMix and Site
seeded.correct <- bind_rows(species.seeded.in.mix,
                            species.seeded.not.in.mix,
                            species.unk.seeded,
                            species.na.seeded,
                            species.no.seeded) %>%
  distinct(.keep_all = TRUE) %>%
  arrange(PlotMix) %>%
  arrange(Name) %>%
  arrange(Site) %>%
  arrange(Region) %>%
  arrange(desc(SpeciesSeeded))

#   Look for duplicates with conflicting SpeciesSeeded info
seeded.correct[duplicated(seeded.correct[, 1:9]), ] # none

# Check possible values of SpeciesSeeded
unique(seeded.correct$SpeciesSeeded) # should be 0, Yes, No

# Look for codes that exist in subplot data but not in seeded.correct
setdiff(unique(subplot$Code), unique(seeded.correct$Code)) # should be 0


# Assign corrected SpeciesSeeded to subplot data
subplot <- subplot %>%
  select(-SpeciesSeeded)
subplot <- left_join(subplot, seeded.correct)

# Check for matching lengths (Sonoran sites only)
nrow(subplot) == (nrow(subplot.cen.raw) + nrow(subplot.se.raw) - 20)

# Look for NAs
unique(subplot$SpeciesSeeded) # should be no NAs


# Save intermediate subplot with correct SeededSpecies,
#   but PlantSource not yet addressed
subplot2 <- subplot



# Add additional grouping columns -----------------------------------------

# PlantSource: Unknown_recruit, Native_recruit, Introduced/Invasive, Seeded, Likely native_recruit, 0
# Weedy: Weedy, Desirable, 0
# PlantSource2: Recruit, Native_recruit, Introduced/Invasive, Seeded, 0
# PlotMix_Climate: None, Current, Projected, 0


## Add PlantSource & PlantSource2 columns ---------------------------------

# Because plants couldn't always be identified to the species level, H. Farrell
#   grouped them by Native_Recruited, Invasive, and Seeded.

unique(subplot$Native)
unique(subplot$SpeciesSeeded)

subplot <- subplot %>%
  mutate(PlantSource = paste0(subplot$Native, "_", subplot$SpeciesSeeded))
unique(subplot$PlantSource)

# Create PlantSource column
subplot <- subplot %>%
  mutate(PlantSource = case_when(
    PlantSource == "0_0" ~ "0",
    PlantSource == "Unknown_No" ~ "Unknown_recruit",
    PlantSource == "Native_No" ~ "Native_recruit",
    PlantSource == "Introduced_No" ~ "Introduced/Invasive",
    PlantSource == "Native_Yes" ~ "Seeded",
    PlantSource == "Native/Unknown_No" ~ "Likely native_recruit",
    PlantSource == "Unknown_Yes" ~ "Seeded"))
unique(subplot$PlantSource)

# Create PlantSource2 column
unique(subplot$PlantSource)
subplot <- subplot %>%
  mutate(PlantSource2 = case_when(
    subplot$PlantSource == "Unknown_recruit" ~ "Recruit",
    str_detect(subplot$PlantSource, "Native_recruit|Likely native_recruit") ~ "Native recruit",
    TRUE ~ subplot$PlantSource))
unique(subplot$PlantSource2)



## Add Weedy column -------------------------------------------------------

# To further simplify weedy vs. desirable species

# Add Weedy column
unique(subplot$PlantSource)
subplot <- subplot %>%
  mutate(Weedy = case_when(
    str_detect(subplot$PlantSource, "Unknown_recruit|Introduced/Invasive") ~ "Weedy",
    str_detect(subplot$PlantSource, "Native_recruit|Likely native_recruit|Seeded") ~ "Desirable",
    TRUE ~ subplot$PlantSource))
unique(subplot$Weedy)



## Add PlotMix_Climate column ---------------------------------------------

# PlotMix names alone cannot be used to group what is climate-adapted and what is current-adapted
#   because mixes are site-specific, and some mixes that are current-adapted are climate-adapted at other sites
#   and vise versa.
# Need ot manually check each site and its seedmix to figure out which is which.

# Add PlotMix_Climate col
subplot <- subplot %>%
  mutate(PlotMix_Climate = case_when(
    str_detect(subplot$Site, "Patagonia|SRER") &
      subplot$PlotMix == "Medium" ~ "Current",
    str_detect(subplot$Site, "Patagonia|SRER") &
      subplot$PlotMix == "Warm" ~ "Projected",
    str_detect(subplot$Site, "Preserve|SCC|Roosevelt|Pleasant") &
      subplot$PlotMix == "Cool" ~ "Current",
    str_detect(subplot$Site, "Preserve|SCC|Roosevelt|Pleasant") &
      subplot$PlotMix == "Warm" ~ "Projected",
    TRUE ~ subplot$PlotMix))
subplot$PlotMix_Climate <- factor(subplot$PlotMix_Climate,
                                  levels = c("None", "Current", "Projected"))


# Save intermediate subplot with all correct plant species info, grouping cols added,
#   but response variables Count and Height not yet addressed
subplot3 <- subplot



# Address Count values of 0 -----------------------------------------------

# Examine rows with Count of 0
count0 <- subplot %>%
  filter(Count == 0)
unique(count0$Code)
#   Code 0 should have a Count of 0 also, but non-0 Codes shouldn't exist

# Examine non-0 Codes
count0.non0code <- count0 %>%
  filter(Code != "0")
#   These rows should just be removed, because we want rows only of plants that were
#     actually observed to be present.

# Remove non-0 Codes from subplot data
subplot <- subplot %>%
  filter(!raw.row %in% count0.non0code$raw.row)



# Address NAs in Count ----------------------------------------------------

# Examine NAs for Count
na.count <- subplot %>%
  filter(is.na(Count))
#   Most rows are for Code 0, in which case Count and Height should also be 0

# Fix Code 0s
na.count.0 <- na.count %>%
  filter(Code == "0")
na.count.0.fix <- na.count.0 %>%
  mutate(Count = 0,
         Height = 0)

# Replace fixed rows in subplot data for 0 Codes
subplot <- subplot %>%
  mutate(Count = as.numeric(Count),
         Height = as.numeric(Height)) %>% 
  filter(!raw.row %in% na.count.0.fix$raw.row) %>%
  bind_rows(na.count.0.fix)


# Examine non-0 Codes
na.count.non0 <- na.count %>%
  filter(Code != "0")

# Original scanned data sheets are not available for this one, so these rows will have to be removed
subplot <- subplot %>%
  filter(!is.na(Count))



# Examine completely empty plots ------------------------------------------

#  Empty subplots have Code 0, and should have only one row (one SiteDatePlotID observation).

# Empty subplots
empty.subplot <- subplot %>%
  filter(Code == "0")
length(unique(empty.subplot$SiteDatePlotID)) == nrow(empty.subplot) # all have only one row (no fix needed)



# Reorder columns ---------------------------------------------------------

subplot <- subplot %>% 
  select(Region, Site, Date_Seeded, Date_Monitored, SiteDateID, Plot, Treatment, PlotMix, 
         PlotMix_Climate, SitePlotID, SiteDatePlotID,
         CodeOriginal, Code, Name, Native, Duration, Lifeform, SpeciesSeeded,
         PlantSource, PlantSource2, Weedy,
         Count, Height, raw.row)


# Write clean subplot data to csv -----------------------------------------

write_csv(subplot,
          file = "Sonoran-data/cleaned/04.1_subplot-data_clean.csv")


save.image("Sonoran-RData/04.1_data-wrangling_subplot.RData")
